+++ data_files/config.json
@@ -1,14 +1,14 @@
 {
-  "no_generations": 100,
-  "starting_population_mode": "from_file",
-  "starting_population_file": "starting_population.json",
-  "breed_function": "random_delete_breed",
-  "mutate_function": "random_delete_mutation",
-  "select_function": "select_children",
-  "alpha": 0.25,
-  "beta": 0.25,
-  "gamma": 0.25,
-  "delta": 100,
+  "no_generations": 50,
+  "starting_population_mode": "auto",
+  "starting_population_size": 50,
+  "breed_function": "dominant_solution_breed_happy_vs_productive",
+  "mutate_function": "delete_lowest_priority",
+  "select_function": "dominant_solution_select",
+  "alpha": 30,
+  "beta": 1,
+  "gamma": 3000,
+  "delta": 20,
   "L": 40,
-  "data_load_mode": "matrices"
+  "data_load_mode": "generated"
 }
